requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
langgraph>=0.0.20
langchain>=0.1.0
langchain-aws>=0.1.0
langchain-community>=0.0.10
boto3>=1.34.0
yfinance>=0.2.0
faiss-cpu>=1.7.4
pypdf>=3.17.0
python-jose[cryptography]>=3.3.0
sentence-transformers>=2.2.2
